# Fix KVM not detecting containerd preload