### Cause
We were doing a `sudo systemctl restart containerd` right before we checked if the preloaded images are present. It was common on KVM that containerd wouldn't finish restarting before running `sudo crictl images --output json` which would fail, and would return `false` for the images being preloaded, resulting in the images be redownloaded unnecessarily.

### Fix
Added a retry to the `sudo crictl images --output json` command so if containerd is still restarting we don't report that the preloaded images aren't present.

### Result
#### First time start
Before: 59s
After: 38.6s
**20.4s speedup (35%)** 

#### Normal start
Before: 48.7s
After: 30.5s
**18.2s speedup (37%)**

Also fixes the `TestErrorSpam/setup` test failure on KVM containerd due to the `❌  Unable to load cached images` that was being output before.

**Before:**
```
$ time minikube start --driver kvm --container-runtime containerd
😄  minikube v1.32.0 on Ubuntu 20.04
✨  Using the kvm2 driver based on user configuration
💾  Downloading driver docker-machine-driver-kvm2:
    > docker-machine-driver-kvm2-...:  65 B / 65 B [---------] 100.00% ? p/s 0s
    > docker-machine-driver-kvm2-...:  13.01 MiB / 13.01 MiB [ 100.00% ? p/s 0s
💿  Downloading VM boot image ...
    > minikube-v1.32.1-1700142131...:  65 B / 65 B [---------] 100.00% ? p/s 0s
    > minikube-v1.32.1-1700142131...:  293.18 MiB / 293.18 MiB  100.00% 240.78 
👍  Starting control plane node minikube in cluster minikube
💾  Downloading Kubernetes v1.28.4 preload ...
    > preloaded-images-k8s-v18-v1...:  436.27 MiB / 436.27 MiB  100.00% 235.22 
🔥  Creating kvm2 VM (CPUs=2, Memory=6000MB, Disk=20000MB) ...
📦  Preparing Kubernetes v1.28.4 on containerd 1.7.9 ...
❌  Unable to load cached images: loading cached images: stat /home/powellsteven/.minikube/cache/images/amd64/registry.k8s.io/kube-proxy_v1.28.4: no such file or directory
    ▪ Generating certificates and keys ...
    ▪ Booting up control plane ...
    ▪ Configuring RBAC rules ...
🔗  Configuring bridge CNI (Container Networking Interface) ...
🔎  Verifying Kubernetes components...
    ▪ Using image gcr.io/k8s-minikube/storage-provisioner:v5
🌟  Enabled addons: storage-provisioner, default-storageclass
🏄  Done! kubectl is now configured to use "minikube" cluster and "default" namespace by default

real	0m59.048s
user	0m4.741s
sys	0m2.471s
```

**After:**
```
$ time minikube start --driver kvm --container-runtime containerd
😄  minikube v1.32.0 on Ubuntu 20.04
✨  Using the kvm2 driver based on user configuration
💾  Downloading driver docker-machine-driver-kvm2:
    > docker-machine-driver-kvm2-...:  65 B / 65 B [---------] 100.00% ? p/s 0s
    > docker-machine-driver-kvm2-...:  13.01 MiB / 13.01 MiB [ 100.00% ? p/s 0s
💿  Downloading VM boot image ...
    > minikube-v1.32.1-1700142131...:  65 B / 65 B [---------] 100.00% ? p/s 0s
    > minikube-v1.32.1-1700142131...:  293.18 MiB / 293.18 MiB  100.00% 283.11 
👍  Starting control plane node minikube in cluster minikube
💾  Downloading Kubernetes v1.28.4 preload ...
    > preloaded-images-k8s-v18-v1...:  436.27 MiB / 436.27 MiB  100.00% 212.95 
🔥  Creating kvm2 VM (CPUs=2, Memory=6000MB, Disk=20000MB) ...
📦  Preparing Kubernetes v1.28.4 on containerd 1.7.9 ...
    ▪ Generating certificates and keys ...
    ▪ Booting up control plane ...
    ▪ Configuring RBAC rules ...
🔗  Configuring bridge CNI (Container Networking Interface) ...
🔎  Verifying Kubernetes components...
    ▪ Using image gcr.io/k8s-minikube/storage-provisioner:v5
🌟  Enabled addons: storage-provisioner, default-storageclass
🏄  Done! kubectl is now configured to use "minikube" cluster and "default" namespace by default

real	0m38.648s
user	0m4.669s
sys	0m2.661s
```